pyproject.toml
[project]
name = "tech-coach-stat"
version = "0.1.0"
description = "Example of how to produce statistics for meteorological data."
authors = [{name = "Arne Sørli", email = "81353974+arneso-ssb@users.noreply.github.com"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "dapla-toolbelt >=3.0.0",
    "dynaconf >=3.2",
    "python-dotenv >=1.0.1",
    "ssb-klass-python >=0.0.11",
    "pandera[mypy] >=0.26.0",
    "isodate >=0.7.2",
    "gcsfs >=2024.12.0",
    "ssb-fagfunksjoner >=1.0.8",
    "ssb-eimerdb>=0.2.4",
    "dapla-auth-client>=1.1.0",
    "pandas (>=2.3.3,<3.0.0)",
    "google-auth>=2.47.0",
    "google-api-core>=2.29.0",
    "requests>=2.32.5",
]

[tool.poetry]
requires-poetry = ">=2.0"
# To include functions in Jupyter Notebooks from other parts of the directory tree,
# without using ProjectRoot from ssb-fagfunksjoner.
packages = [
    { include = "config" },
    { include = "functions", from = "src" },
]

[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = ">=24.4.2"}
ipykernel = ">=6.29"
mypy = ">=1.13.0"
pandas-stubs = ">=2.3.3.260113"
pre-commit = ">=3.3.1"
pre-commit-hooks = ">=4.4.0"
pytest = ">=7.1.3"
pytest-cov = ">=5.0.0"
pytest-mock = ">=3.14.0"
ruff = ">=0.0.284"
types-requests = ">=2.31.0.20240406"
deptry = ">=0.24.0"

[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"

[tool.jupytext]
formats = "ipynb,auto:percent"
notebook_metadata_filter = "jupytext.text_representation,-jupytext.text_representation.jupytext_version,-widgets,-varInspector"
cell_metadata_filter = "-papermill,tags"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_context = true
disallow_any_generics = false

[[tool.mypy.overrides]]
module = [
    "dynaconf.*",
    "gcsfs.*",
    "isodate.*",
]
ignore_missing_imports = true

[tool.ruff]
force-exclude = true  # Apply excludes to pre-commit
show-fixes = true
src = ["src", "tests"]
target-version = "py310"  # Minimum Python version supported
include = ["*.py", "*.pyi", "**/pyproject.toml", "*.ipynb"]
extend-exclude = [
    "__pycache__",
    ".ipynb_checkpoints",
]

# Ruff rules may be customized as desired: https://docs.astral.sh/ruff/rules/
[tool.ruff.lint]
select = [
    "A",    # prevent using keywords that clobber python builtins
    "ANN",  # check type annotations
    "B",    # bugbear: security warnings
    "D",    # documentation
    "E",    # pycodestyle
    "F",    # pyflakes
    "ISC",  # implicit string concatenation
    "I",    # sort imports
    "UP",   # alert you when better syntax is available in your python version
    "RUF",  # the ruff developer's own rules
]
ignore = [
    "ANN202", # Don't requiere return type annotation for private functions.
    "ANN401", # Allow type annotation with type Any.
    "D100",   # Supress undocumented-public-module. Only doc of public api required.
    "E402",   # Supress module-import-not-at-top-of-file, needed in jupyter notebooks.
    "E501",   # Supress line-too-long warnings: trust black's judgement on this one.
]

[tool.ruff.lint.isort]
force-single-line = true
known-first-party = ["config"]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["classmethod", "validator", "root_validator", "pydantic.validator"]

[tool.ruff.lint.per-file-ignores]
"*/__init__.py" = ["F401"]
"**/tests/*" = [
    "ANN001",  # type annotations don't add value for test functions
    "ANN002",  # type annotations don't add value for test functions
    "ANN003",  # type annotations don't add value for test functions
    "ANN201",  # type annotations don't add value for test functions
    "ANN204",  # type annotations don't add value for test functions
    "ANN205",  # type annotations don't add value for test functions
    "ANN206",  # type annotations don't add value for test functions
    "D100",    # docstrings are overkill for test functions
    "D101",
    "D102",
    "D103",
    "S101",    # asserts are encouraged in pytest
]
